cleaning controllers

+++ src/main/java/com/etoiledespoir/onlinekvshop/controller/itemController/categories/hair/HairController.java
@@ -2,10 +2,16 @@ package com.etoiledespoir.onlinekvshop.controller.itemController.categories.hair
 
 import com.etoiledespoir.onlinekvshop.controller.Icontroller;
 import com.etoiledespoir.onlinekvshop.domain.item.impl.Hair;
+import com.etoiledespoir.onlinekvshop.domain.pic.picHelper.MypicHelpRead;
+import com.etoiledespoir.onlinekvshop.factory.domain.item.HairFactory;
+import com.etoiledespoir.onlinekvshop.factory.domain.pic.pictureHelpReader.MypicHelpReadFactory;
 import com.etoiledespoir.onlinekvshop.service.itemService.category.hair.impl.HairService;
+import com.etoiledespoir.onlinekvshop.service.mypic.impl.PictureService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
 
+import java.io.IOException;
 import java.util.List;
 
 @RestController
@@ -13,17 +19,47 @@ import java.util.List;
 public class HairController implements Icontroller<Hair, String> {
 @Autowired
     HairService hairService;
+    @Autowired
+    PictureService pictureService;
+    private boolean resultBool=false;
 
 @PostMapping("/create")
+
+    public Hair createWithFile(@RequestParam("file") MultipartFile file,@RequestParam("name")String name,@RequestParam("type")String type,@RequestParam("size")String size,@RequestParam("desc")String desc,@RequestParam("color")String color) throws IOException {
+    Hair hair= HairFactory.getHair(name,type,size,desc,color);
+    if(hair!=null) {
+        Hair result= hairService.creat(hair);
+        if(result!=null){
+            resultBool= pictureService.helpCreateFile(file,result.getItemNumber());
+            if(resultBool==true){
+                return result;
+            }
+        }
+    }return null;
+    }
+
     @Override
-    public Hair create(@RequestBody Hair hair) {
-        return hairService.creat(hair);
+    public Hair create(Hair hair) {
+        return null;
+    }
+
+    @Override
+    public Hair read(String id) {
+        return null;
     }
 
     @GetMapping("/read")
-    @Override
-    public Hair read(@RequestParam("id") String id) {
-        return hairService.read(id);
+
+    public MypicHelpRead readWithImage(@RequestParam("id") String id) {
+       Hair hair=  hairService.read(id);
+       if(hair!=null){
+           String StringImage = pictureService.readFile(id);
+           if(StringImage!=null||StringImage.equals("")){
+               MypicHelpRead mypicHelpRead= MypicHelpReadFactory.getMypicture(id,hair.getItemNumber(),StringImage,hair.toString());
+                return mypicHelpRead;
+           }
+       }
+       return null;
     }
 
     @PostMapping("/update")
@@ -35,7 +71,16 @@ public class HairController implements Icontroller<Hair, String> {
     @GetMapping("/delete")
     @Override
     public Hair delete(@RequestParam("id") String id) {
-        return hairService.delete(id);
+        Hair hair=null;
+    String readId=readWithImage(id).getId();
+        if(readId!=null){
+             hair= hairService.delete(id);
+
+                if(pictureService.deleteFromFile(id)==true){
+                    return hair;
+                }
+
+        }return null;
     }
 
     @GetMapping("/reads")
