+++ src/main/java/com/etoiledespoir/onlinekvshop/controller/itemController/categories/shoes/ShoesController.java
@@ -2,10 +2,16 @@ package com.etoiledespoir.onlinekvshop.controller.itemController.categories.shoe
 
 import com.etoiledespoir.onlinekvshop.controller.Icontroller;
 import com.etoiledespoir.onlinekvshop.domain.item.impl.Shoes;
+import com.etoiledespoir.onlinekvshop.domain.pic.picHelper.MypicHelpRead;
+import com.etoiledespoir.onlinekvshop.factory.domain.item.ShoesFactory;
+import com.etoiledespoir.onlinekvshop.factory.domain.pic.pictureHelpReader.MypicHelpReadFactory;
 import com.etoiledespoir.onlinekvshop.service.itemService.category.shoes.impl.ShoesService;
+import com.etoiledespoir.onlinekvshop.service.mypic.impl.PictureService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
 
+import java.io.IOException;
 import java.util.List;
 
 @RestController
@@ -14,17 +20,44 @@ public class ShoesController implements Icontroller<Shoes, String> {
 
     @Autowired
     ShoesService shoesService;
+    @Autowired
+    PictureService pictureService;
+    private boolean result=false;
 
     @PostMapping("/create")
+    public Shoes createWithFile(@RequestParam("file") MultipartFile file,@RequestParam("itemName") String marque,@RequestParam("type") String itemType,@RequestParam("size") String size, String gender,@RequestParam("desc") String desc,@RequestParam("clore") String color) throws IOException {
+        Shoes shoes= ShoesFactory.getShoes(marque,size,gender,desc);
+        Shoes result1=shoesService.creat(shoes);
+        if(result1!=null){
+            result= pictureService.helpCreateFile(file,result1.getItemNumber());
+            if(result==true){
+                return result1;
+            }
+        }
+        return null;
+    }
+
     @Override
-    public Shoes create(@RequestBody Shoes shoes) {
-        return shoesService.creat(shoes);
+    public Shoes create(Shoes shoes) {
+        return null;
+    }
+
+    @Override
+    public Shoes read(String id) {
+        return null;
     }
 
     @GetMapping("/read")
-    @Override
-    public Shoes read(@RequestParam("id") String id) {
-        return shoesService.read(id);
+    public MypicHelpRead readWithImage(@RequestParam("id") String id) {
+        Shoes shoes=shoesService.read(id);
+        if(shoes!=null){
+            String StringImage=pictureService.readFile(id);
+            if(StringImage!=null||StringImage.equals("")){
+                MypicHelpRead mypicHelpRead= MypicHelpReadFactory.getMypicture(id,shoes.getItemNumber(),shoes.getColor(),shoes.toString());
+                return mypicHelpRead;
+            }
+        }
+        return null;
     }
 
     @PostMapping("/update")
@@ -36,7 +69,16 @@ public class ShoesController implements Icontroller<Shoes, String> {
     @GetMapping("/delete")
     @Override
     public Shoes delete(@RequestParam("id") String id) {
-        return shoesService.delete(id);
+        Shoes shoes=null;
+        MypicHelpRead result=readWithImage(id);
+        if(result!=null){
+            shoes=shoesService.delete(id);
+            if(pictureService.deleteFromFile(id)==true){
+                return shoes;
+            }
+
+        }
+        return null;
     }
 
     @GetMapping("/reads")
